Create MotionAnimations.
