+++ portfolio/src/components/Home.tsx
@@ -8,18 +8,7 @@ import {
   FaUniversity,
 } from "react-icons/fa";
 import AboutCard from "./AboutCard";
-
-const fadeInUp = {
-  initial: { opacity: 0, y: -50 },
-  animate: { opacity: 1, y: 0 },
-  transition: { duration: 1, delay: 0.5 },
-};
-
-const fadeIn = {
-  initial: { opacity: 0 },
-  animate: { opacity: 1 },
-  transition: { duration: 1, delay: 1 },
-};
+import { fadeInUp, fadeIn } from "../utils/MotionAnimations";
 
 export const Home = () => {
   return (
