+++ portfolio/src/utils/MotionAnimations.tsx
@@ -1,0 +1,13 @@
+const fadeInUp = {
+  initial: { opacity: 0, y: -50 },
+  animate: { opacity: 1, y: 0 },
+  transition: { duration: 1, delay: 0.5 },
+};
+
+const fadeIn = {
+  initial: { opacity: 0 },
+  animate: { opacity: 1 },
+  transition: { duration: 1, delay: 1 },
+};
+
+export { fadeInUp, fadeIn };
